refactor(quil-node): extract message_loop (Batch 4, Step 17)
- master_node/message_loop.rs: Phase 6 message-receive loop —
  spawn(args), takes a MessageLoopArgs struct with 31 fields covering
  the captured stores, OnceLocks, broadcast/loopback channels, peer
  caches, signers, etc. ~764 lines extracted (the entire
  tokio::spawn(async move { loop { tokio::select! { ... } } }) block
  including the top-level GLOBAL_* dispatch match). The message_router
  construction also moves inside spawn (it was only used there).

main.rs: 3,874 → 3,075 lines (~800 lines extracted in one step).
Total reduction from original 6,208: 3,133 lines (50.5%).

Step 18 (further per-bitmask handler split into message_handlers.rs)
deferred — the deep nesting now lives inside master_node/message_loop.rs
which is independently scannable; main.rs's run_master_node body is
already the planned ~15 sequential init/spawn calls.

Verification: cargo check --workspace, cargo nextest run --workspace
(2,326 passed / 13 skipped), simtest/ci.sh -short — all green.
