+++ crates/quil-node/src/main.rs
@@ -338,10 +338,10 @@ async fn run_master_node(
 
     let master_node::networking::P2pHandles {
         p2p_handle,
-        mut msg_rx,
+        msg_rx,
         peer_id,
         consensus_loopback_tx,
-        mut consensus_loopback_rx,
+        consensus_loopback_rx,
         listen_addr,
     } = master_node::networking::init(config, config_dir, network, archive_mode).await?;
 
@@ -510,16 +510,6 @@ async fn run_master_node(
     // ---------------------------------------------------------------
     info!(archive = archive_mode, "master node initialized — waiting for frames");
 
-    let clock_store_recv = clock_store.clone();
-    let exec_mgr_for_recv = exec_manager.clone();
-    let recv_token = token.clone();
-
-    // Global bitmasks for BlossomSub topic subscriptions.
-    const GLOBAL_CONSENSUS: &[u8] = &[0x00];
-    const GLOBAL_FRAME: &[u8] = &[0x00, 0x00];
-    const GLOBAL_PROVER: &[u8] = &[0x00, 0x00, 0x00];
-    const GLOBAL_PEER_INFO: &[u8] = &[0x00, 0x00, 0x00, 0x00];
-    const GLOBAL_ALERT: &[u8] = &[0u8; 16];
 
     // Resolve our Ed448 seed (57 bytes) for the mTLS cert. The peer key in
     // config is either 57 bytes (raw seed) or 114 bytes (seed + pubkey).
@@ -1810,842 +1800,39 @@ async fn run_master_node(
         quil_rpc::global_service::GLOBAL_MESSAGE_BROADCAST_CAPACITY,
     )
     .0;
-    let gmtx_for_recv = global_msg_tx.clone();
-
-    let pool_for_recv = archive_pool.clone();
-    let mtls_seed_for_recv: Option<[u8; 57]> = mtls_seed;
-    let hg_store_for_recv = hg_store.clone();
-    let frame_validator_for_recv = frame_validator;
-    let mc_for_recv = message_collector.clone();
-    let coverage_for_recv = coverage_monitor.clone();
-    let wa_for_recv = worker_allocator.clone();
-    let pp_for_recv = prover_pipeline.clone();
-    let ch_for_recv = consensus_handle.clone();
-    let va_for_recv = vote_aggregator.clone();
-    let ta_for_recv = timeout_aggregator.clone();
-    let pic_for_recv = peer_info_cache.clone();
-    let shard_engines_for_recv = shard_engines.clone();
-    let sr_for_recv = signer_registry.clone();
-    let cf_for_recv = current_frame.clone();
-    let lhf_for_recv = last_global_head_frame.clone();
-    let genesis_archive_peer_ids_for_recv = genesis_archive_peer_ids.clone();
-    let genesis_prover_addrs_for_recv = genesis_prover_addrs.clone();
-    let alert_pubkey_for_recv: Vec<u8> = hex::decode(&config.engine.alert_key).unwrap_or_default();
-    // Mainnet validates archive-capability claims against the
-    // hardcoded genesis archive peer IDs. Testnet/devnet bootstraps
-    // can't use that list (those keys aren't theirs), so the receive
-    // loop accepts any archive-claiming peer when network != 0.
-    // Use the CLI-supplied network value (`--network`); the YAML's
-    // `p2p.network` field is left at 0 in shared configs.
-    let network_for_recv: u8 = network;
-    let archive_mode_for_recv: bool = archive_mode;
-    let pl_for_recv = prover_lifecycle.clone();
-    let pr_for_recv = prover_registry.clone();
-    let wm_for_recv = worker_manager.clone();
-    let archive_mode_recv = archive_mode;
-    let reward_issuer = Arc::new(quil_engine::OptRewardIssuance);
-    let pa_for_recv = prover_address;
-    let p2p_for_recv = p2p_handle.clone();
-
-    // Per-bitmask validator gate. Malformed bytes are dropped here so
-    // the dispatch loop below stays cheap. Topics without a registered
-    // validator fall through unchanged.
-    let message_router = Arc::new(quil_engine::message_router::MessageRouter::new());
-    message_router.register_validator(
-        GLOBAL_PEER_INFO.to_vec(),
-        quil_engine::message_router::validator_global_peer_info(),
-    );
-    message_router.register_validator(
-        GLOBAL_PROVER.to_vec(),
-        quil_engine::message_router::validator_global_prover(),
-    );
-    message_router.register_validator(
-        GLOBAL_FRAME.to_vec(),
-        quil_engine::message_router::validator_global_frame(),
-    );
-    message_router.register_validator(
-        GLOBAL_CONSENSUS.to_vec(),
-        quil_engine::message_router::validator_global_consensus(),
-    );
-    let router_for_recv = message_router.clone();
-
-    tokio::spawn(async move {
-        let mut frames_received: u64 = 0;
-        let mut peer_infos_received: u64 = 0;
-        let mut peer_info_digest_cache: std::collections::HashSet<[u8; 32]> = std::collections::HashSet::new();
-        let mut archive_peers_seen: std::collections::HashSet<Vec<u8>> = std::collections::HashSet::new();
-        let mut consensus_msgs_received: u64 = 0;
-        let mut prover_msgs_received: u64 = 0;
-        let mut router_drops: u64 = 0;
-        let mut status_timer = tokio::time::interval(std::time::Duration::from_secs(30));
-        status_timer.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-        // Track the highest frame number we've fully executed (through
-        // the execution manager + lifecycle). New frames arriving via
-        // gossip can be wildly out of order; only execute consecutive
-        // frames starting from `last_executed + 1` so prover-state
-        // dependent ops (ProverConfirm/Resume against a prover the
-        // previous frame registered) don't fail spuriously. Seeded
-        // from the latest frame already in the clock store so a
-        // restart picks up where it left off.
-        let mut last_executed_frame: u64 = clock_store_recv
-            .get_latest_global_frame()
-            .ok()
-            .and_then(|f| f.header.as_ref().map(|h| h.frame_number))
-            .unwrap_or(0);
-        loop {
-            tokio::select! {
-                _ = status_timer.tick() => {
-                    // Periodic allocation status snapshot.
-                    let peer_count = p2p_for_recv.peer_count();
-                    let latest_frame = clock_store_recv.get_latest_global_frame()
-                        .ok()
-                        .and_then(|f| f.header.as_ref().map(|h| h.frame_number))
-                        .unwrap_or(0);
-                    let (active, pending, total_allocs) = {
-                        use quil_types::consensus::{EffectiveStatus, ProverRegistry};
-                        match pr_for_recv.get_prover_info(&pa_for_recv) {
-                            Ok(Some(info)) => {
-                                let mut a = 0usize;
-                                let mut p = 0usize;
-                                let mut total = 0usize;
-                                for alloc in &info.allocations {
-                                    match alloc.effective_status(latest_frame) {
-                                        EffectiveStatus::Active => {
-                                            a += 1;
-                                            total += 1;
-                                        }
-                                        EffectiveStatus::Joining => {
-                                            p += 1;
-                                            total += 1;
-                                        }
-                                        EffectiveStatus::Paused | EffectiveStatus::Leaving => {
-                                            total += 1;
-                                        }
-                                        _ => {}
-                                    }
-                                }
-                                (a, p, total)
-                            }
-                            _ => (0, 0, 0),
-                        }
-                    };
-                    info!(
-                        peers = peer_count,
-                        frame = latest_frame,
-                        frames_received,
-                        active_shards = active,
-                        pending_joins = pending,
-                        total_allocations = total_allocs,
-                        peer_infos = peer_infos_received,
-                        archive_peers = archive_peers_seen.len(),
-                        consensus_msgs = consensus_msgs_received,
-                        prover_msgs = prover_msgs_received,
-                        router_drops,
-                        "node status"
-                    );
-                }
-                msg = async {
-                    // Merge the network receive channel and the
-                    // self-loopback channel — both produce
-                    // `ReceivedMessage`s that go through the same
-                    // dispatch logic. This is how the proposer's own
-                    // proposal reaches its own `vote_aggregator` and
-                    // event_loop without relying on BlossomSub
-                    // self-echo (which doesn't happen).
-                    tokio::select! {
-                        biased;
-                        m = consensus_loopback_rx.recv() => m,
-                        m = msg_rx.recv() => m,
-                    }
-                } => {
-                    match msg {
-                        Some(received) => {
-                            // Forward to connected StreamGlobalMessages
-                            // subscribers (workers) — ONLY peer-info.
-                            //
-                            // In Go, `broadcastGlobalMessage` is called
-                            // from GLOBAL_FRAME, GLOBAL_PROVER, and
-                            // GLOBAL_PEER_INFO handlers. But on a
-                            // non-archive master those handlers don't
-                            // fire (not subscribed). The only messages
-                            // arriving here are GLOBAL_PEER_INFO (from
-                            // the mesh) and GLOBAL_CONSENSUS (from the
-                            // self-loopback when the master produces
-                            // its own proposals/votes). Workers don't
-                            // need either — they get global frames
-                            // from the archive poller, prover messages
-                            // via direct gRPC, and per-shard consensus
-                            // from their own mesh subscriptions.
-                            // Forwarding the loopback's
-                            // GLOBAL_CONSENSUS was the source of QC
-                            // verification failures (all-zero sig/pk,
-                            // 0xFF bitmask from genesis QC data) and
-                            // OOM (on archive masters, the [0xFF;32]
-                            // catch-all subscription fed every shard's
-                            // traffic into the broadcast channel).
-                            //
-                            // The only bitmask workers still need from
-                            // the master stream is GLOBAL_PEER_INFO
-                            // (peer discovery for their own mesh).
-                            if received.bitmask.as_slice() == GLOBAL_PEER_INFO {
-                                let _ = gmtx_for_recv.send(
-                                    quil_types::proto::global::StreamGlobalMessagesResponse {
-                                        data: received.data.clone(),
-                                        bitmask: received.bitmask.clone(),
-                                    },
-                                );
-                            }
-
-                            // Per-topic validator gate. Malformed bytes are
-                            // dropped before they reach a queue.
-                            // Unregistered topics fall through.
-                            let route_outcome = router_for_recv
-                                .route(&received.bitmask, &received.data);
-                            if !route_outcome.should_dispatch() {
-                                router_drops += 1;
-                                // Categorize for operator visibility.
-                                let topic = match received.bitmask.as_slice() {
-                                    GLOBAL_PEER_INFO => "peer_info",
-                                    GLOBAL_PROVER => "prover",
-                                    GLOBAL_FRAME => "frame",
-                                    GLOBAL_CONSENSUS => "consensus",
-                                    GLOBAL_ALERT => "alert",
-                                    _ => "shard/unknown",
-                                };
-                                let type_prefix = if received.data.len() >= 4 {
-                                    format!("0x{:08x}", u32::from_be_bytes(
-                                        received.data[..4].try_into().unwrap_or([0;4])
-                                    ))
-                                } else {
-                                    format!("short({}B)", received.data.len())
-                                };
-                                debug!(
-                                    topic,
-                                    type_prefix,
-                                    len = received.data.len(),
-                                    total_dropped = router_drops,
-                                    "router validator dropped message",
-                                );
-                                continue;
-                            }
-
-                            match received.bitmask.as_slice() {
-                            GLOBAL_PEER_INFO => {
-                                match quil_p2p::classify_peer_info_message(&received.data) {
-                                    Ok(quil_p2p::PeerInfoMessage::PeerInfo(info)) => {
-                                        peer_infos_received += 1;
-                                        // Dedup: hash PeerInfo with timestamp zeroed
-                                        // (mirrors Go's hashPeerInfo). Skip if seen.
-                                        let mut dedup_info = info.clone();
-                                        dedup_info.timestamp = 0;
-                                        let dedup_payload = quil_p2p::encode_canonical_peer_info(
-                                            &dedup_info,
-                                            &dedup_info.public_key,
-                                            &dedup_info.signature,
-                                        );
-                                        use sha2::Digest as _;
-                                        let digest: [u8; 32] = sha2::Sha256::digest(&dedup_payload).into();
-                                        if !peer_info_digest_cache.insert(digest) {
-                                            continue;
-                                        }
-                                        if peer_info_digest_cache.len() > 10_000 {
-                                            peer_info_digest_cache.clear();
-                                        }
-                                        if !info.peer_id.is_empty() {
-                                            let mut cache = pic_for_recv.write();
-                                            cache.insert(info.peer_id.clone(), info.clone());
-                                        }
-                                        // Only ARCHIVE-capable peers go into the
-                                        // poll pool. Plain peers reject every
-                                        // GetGlobalFrame call with "not currently
-                                        // syncable".
-                                        if info.is_archive() {
-                                            // Validate against genesis archive peers.
-                                            // The peer_id in PeerInfo is raw bytes;
-                                            // genesis has base58 peer IDs. Convert
-                                            // PeerInfo peer_id to hex for comparison
-                                            // against genesis BLS pubkey hashes.
-                                            let peer_hex = bs58::encode(&info.peer_id).into_string();
-                                            // On testnet/devnet (network != 0) the genesis
-                                            // archive list isn't ours, so we accept any
-                                            // archive-claiming peer. Mainnet keeps the
-                                            // strict allowlist check below.
-                                            let is_genesis_archive = network_for_recv != 0
-                                                || genesis_archive_peer_ids_for_recv
-                                                    .contains(&info.peer_id);
-                                            if !is_genesis_archive {
-                                                warn!(
-                                                    peer = peer_hex,
-                                                    from = bs58::encode(&received.from).into_string(),
-                                                    "FAKE ARCHIVE — peer claims archive capability but is not a genesis archive peer"
-                                                );
-                                                continue;
-                                            }
-                                            let is_new = archive_peers_seen.insert(info.peer_id.clone());
-                                            if is_new {
-                                                info!(
-                                                    peer = peer_hex,
-                                                    head_frame = info.last_global_head_frame,
-                                                    total = archive_peers_seen.len(),
-                                                    "verified genesis archive peer"
-                                                );
-                                            }
-                                            let mut first_addr: Option<String> = None;
-                                            for reach in &info.reachability {
-                                                for ma in &reach.stream_multiaddrs {
-                                                    if let Some(addr) = crate::util::multiaddr::multiaddr_to_host_port_with_network(ma, network_for_recv) {
-                                                        if first_addr.is_none() {
-                                                            first_addr = Some(addr.clone());
-                                                        }
-                                                        pool_for_recv.add(addr).await;
-                                                    }
-                                                }
-                                            }
-                                            info!(
-                                                peer = bs58::encode(&info.peer_id).into_string(),
-                                                head_frame = info.last_global_head_frame,
-                                                total_archives = archive_peers_seen.len(),
-                                                "discovered archive peer"
-                                            );
-                                            // First archive: sync all four
-                                            // CRDT phases of the global
-                                            // prover tree sequentially. Each
-                                            // ensure_prover_tree call either
-                                            // loads the cached blob from
-                                            // RocksDB or pulls + verifies +
-                                            // persists from this archive.
-                                            // Skip in archive mode — archives
-                                            // have full history locally and
-                                            // the legacy whole-tree blob
-                                            // sync path isn't populated by
-                                            // migrated stores (per-vertex
-                                            // data at `0x30` is, but blob
-                                            // cache at `0x2F` is not), so
-                                            // calling this against another
-                                            // archive that's also fresh
-                                            // from migration just trades
-                                            // "no tree data available"
-                                            // errors back and forth.
-                                            if is_new && archive_peers_seen.len() == 1
-                                                && !archive_mode_for_recv {
-                                                if let (Some(seed), Some(addr)) =
-                                                    (mtls_seed_for_recv, first_addr)
-                                                {
-                                                    let store = hg_store_for_recv.clone();
-                                                    let cs = clock_store_recv.clone();
-                                                    tokio::spawn(async move {
-                                                        use quil_types::proto::application::HypergraphPhaseSet::*;
-                                                        // Pin sync against the most-recent verified
-                                                        // frame's prover_tree_commitment (when
-                                                        // available). Empty during bootstrap before
-                                                        // any frame is stored.
-                                                        let expected_root = cs
-                                                            .get_latest_global_frame()
-                                                            .ok()
-                                                            .and_then(|f| f.header.map(|h| h.prover_tree_commitment))
-                                                            .unwrap_or_default();
-                                                        for phase in [VertexAdds, VertexRemoves, HyperedgeAdds, HyperedgeRemoves] {
-                                                            match quil_rpc::ensure_prover_tree(
-                                                                &addr,
-                                                                &seed,
-                                                                phase,
-                                                                store.clone(),
-                                                                &expected_root,
-                                                            ).await {
-                                                                Ok(stats) => {
-                                                                    info!(
-                                                                        addr = %addr,
-                                                                        ?phase,
-                                                                        matched = stats.commitments_match,
-                                                                        leaves = stats.leaves_pulled,
-                                                                        "phase sync complete"
-                                                                    );
-                                                                }
-                                                                Err(e) => {
-                                                                    warn!(addr = %addr, ?phase, error = %e, "ensure_prover_tree failed");
-                                                                    break;
-                                                                }
-                                                            }
-                                                        }
-                                                        info!("all 4 phases synced");
-
-                                                        // Build the in-memory ProverRegistry
-                                                        // from the persisted vertex store.
-                                                        let mut registry =
-                                                            quil_execution::InMemoryProverRegistry::new();
-                                                        registry.refresh(&store);
-                                                        info!(
-                                                            provers_visited = registry.provers_visited(),
-                                                            allocations_visited = registry.allocations_visited(),
-                                                            rewards_visited = registry.rewards_visited(),
-                                                            distinct_provers = registry.distinct_provers(),
-                                                            distinct_filters = registry.distinct_filters(),
-                                                            "prover registry refreshed"
-                                                        );
-
-                                                        // Sample a few active provers.
-                                                        let all_active =
-                                                            registry.get_all_active_app_shard_provers();
-                                                        info!(
-                                                            active_count = all_active.len(),
-                                                            "active prover count from registry"
-                                                        );
-                                                        for prover in all_active.iter().take(3) {
-                                                            info!(
-                                                                address = %hex::encode(&prover.address),
-                                                                seniority = prover.seniority,
-                                                                available_storage = prover.available_storage,
-                                                                allocations = prover.allocations.len(),
-                                                                "  active prover"
-                                                            );
-                                                        }
-                                                    });
-                                                }
-                                            }
-                                        } else if peer_infos_received <= 5
-                                            || peer_infos_received % 100 == 0
-                                        {
-                                            info!(
-                                                total_peer_infos = peer_infos_received,
-                                                total_archives = archive_peers_seen.len(),
-                                                "PeerInfo discovery progress"
-                                            );
-                                        }
-                                    }
-                                    Ok(quil_p2p::PeerInfoMessage::KeyRegistry) => {
-                                        // Decode and stash in the signer registry so
-                                        // consensus-message BLS signatures from the
-                                        // announcing peer can later be verified using
-                                        // the prover key bound to its Ed448 identity.
-                                        // Older-timestamp replays are ignored inside
-                                        // `SignerRegistry::update`.
-                                        match quil_p2p::decode_canonical_key_registry(&received.data) {
-                                            Ok(reg) => {
-                                                let identity_len = reg.ed448_pubkey.len();
-                                                let prover_len = reg.bls_pubkey.len();
-                                                sr_for_recv.update(reg);
-                                                debug!(
-                                                    identity_len,
-                                                    prover_len,
-                                                    total_entries = sr_for_recv.len(),
-                                                    "ingested KeyRegistry"
-                                                );
-                                            }
-                                            Err(e) => {
-                                                warn!(error = %e, "failed to decode KeyRegistry");
-                                            }
-                                        }
-                                    }
-                                    Ok(quil_p2p::PeerInfoMessage::Unknown(prefix)) => {
-                                        warn!(prefix = format!("0x{:04x}", prefix),
-                                            "unknown PEER_INFO bitmask message type");
-                                    }
-                                    Err(e) => {
-                                        warn!(error = %e, "failed to decode PeerInfo");
-                                    }
-                                }
-                            }
-                            GLOBAL_FRAME => {
-                                // Try canonical bytes first (the wire format),
-                                // fall back to proto decode (archive poller uses proto).
-                                let frame_result: std::result::Result<quil_types::proto::global::GlobalFrame, _> =
-                                    quil_engine::consensus_wire::decode_global_frame(&received.data)
-                                        .or_else(|canonical_err| {
-                                            warn!(error = %canonical_err, "canonical decode failed, trying proto");
-                                            prost::Message::decode(received.data.as_slice())
-                                                .map_err(|e| quil_types::error::QuilError::InvalidArgument(
-                                                    format!("failed to decode Protobuf message: {} (canonical: {})", e, canonical_err)
-                                                ))
-                                        });
-                                match frame_result {
-                                    Ok(frame) => {
-                                        let frame_num = frame.header.as_ref()
-                                            .map(|h| h.frame_number).unwrap_or(0);
-
-                                        // Validate prover is a genesis prover
-                                        if let Some(h) = frame.header.as_ref() {
-                                            if !genesis_prover_addrs_for_recv.contains(&h.prover) {
-                                                warn!(
-                                                    frame = frame_num,
-                                                    prover = hex::encode(&h.prover),
-                                                    from = bs58::encode(&received.from).into_string(),
-                                                    "INVALID PROVER — not a genesis prover, possible attacker"
-                                                );
-                                                continue;
-                                            }
-                                        }
-
-                                        // Verify VDF proof before storing.
-                                        // Wrap in catch_unwind — the classgroup can panic
-                                        // on malformed VDF output from canonical decode bugs.
-                                        let validate_result = std::panic::catch_unwind(
-                                            std::panic::AssertUnwindSafe(|| frame_validator_for_recv.validate(&frame))
-                                        );
-                                        match validate_result {
-                                            Ok(Ok(true)) => {}
-                                            Ok(Ok(false)) => {
-                                                // Validator returned false — either VDF or BLS
-                                                // signature check rejected it. The specific
-                                                // reason is logged by `GlobalFrameVerifier::validate`.
-                                                warn!(frame = frame_num, "frame rejected by validator — dropping");
-                                                continue;
-                                            }
-                                            Ok(Err(e)) => {
-                                                warn!(frame = frame_num, error = %e, "VDF validation error — dropping frame");
-                                                continue;
-                                            }
-                                            Err(_) => {
-                                                warn!(
-                                                    frame = frame_num,
-                                                    output_len = frame.header.as_ref().map(|h| h.output.len()).unwrap_or(0),
-                                                    "VDF validation PANIC — frame output likely corrupted, dropping"
-                                                );
-                                                continue;
-                                            }
-                                        }
-
-                                        match clock_store_recv.put_global_frame(&frame, None) {
-                                            Ok(()) => {
-                                                frames_received += 1;
-                                                // `observe` / `fetch_max` never
-                                                // regress these counters below
-                                                // an already-seen value (e.g.
-                                                // if a stale duplicate frame
-                                                // arrives out-of-order via
-                                                // BlossomSub).
-                                                cf_for_recv.observe(frame_num);
-                                                lhf_for_recv.fetch_max(frame_num, std::sync::atomic::Ordering::Relaxed);
-
-                                                // Frame execution dispatches on node mode:
-                                                //
-                                                //   * Archive nodes need contiguous frame
-                                                //     history (clients sync from us; gaps
-                                                //     break their replay). We drain the
-                                                //     clock store in order from
-                                                //     `last_executed + 1` forward and
-                                                //     wait for missing predecessors.
-                                                //
-                                                //   * Non-archive nodes only need to
-                                                //     follow the chain head. BlossomSub
-                                                //     gossip is unordered; stale / backfill
-                                                //     frames are noise. We skip any frame
-                                                //     whose number is not strictly newer
-                                                //     than what we've already executed,
-                                                //     and never wait for missing
-                                                //     predecessors — they're already past.
-                                                let frames_to_execute: Vec<(u64, quil_types::proto::global::GlobalFrame)> =
-                                                if archive_mode_recv {
-                                                    let mut out = Vec::new();
-                                                    loop {
-                                                        let next_num = last_executed_frame.saturating_add(1);
-                                                        match clock_store_recv.get_global_frame(next_num) {
-                                                            Ok(f) => out.push((next_num, f)),
-                                                            Err(_) => break,
-                                                        }
-                                                        last_executed_frame = next_num;
-                                                    }
-                                                    if frame_num > last_executed_frame {
-                                                        debug!(
-                                                            frame = frame_num,
-                                                            awaiting = last_executed_frame + 1,
-                                                            "archive: stored out-of-order frame, awaiting predecessor"
-                                                        );
-                                                    }
-                                                    out
-                                                } else if frame_num > last_executed_frame {
-                                                    last_executed_frame = frame_num;
-                                                    vec![(frame_num, frame.clone())]
-                                                } else {
-                                                    debug!(
-                                                        frame = frame_num,
-                                                        last_executed = last_executed_frame,
-                                                        "non-archive: skipping stale/backfill frame",
-                                                    );
-                                                    Vec::new()
-                                                };
-
-                                                for (exec_num, exec_frame) in frames_to_execute {
-                                                    match quil_engine::frame_processor::process_global_frame_with_rewards(
-                                                        &exec_mgr_for_recv,
-                                                        &exec_frame,
-                                                        &num_bigint::BigInt::from(1),
-                                                        Some(reward_issuer.as_ref() as &dyn quil_types::consensus::RewardIssuance),
-                                                        Some(pr_for_recv.as_ref() as &dyn quil_types::consensus::ProverRegistry),
-                                                    ) {
-                                                        Ok((applied, skipped)) => {
-                                                            info!(
-                                                                frame = exec_num,
-                                                                total = frames_received,
-                                                                applied,
-                                                                skipped,
-                                                                "received + processed GlobalFrame"
-                                                            );
-                                                            coverage_for_recv.check(exec_num);
-                                                            if !archive_mode_recv {
-                                                                if let Err(e) = wa_for_recv.on_new_frame(exec_num) {
-                                                                    warn!(error = %e, "worker allocation failed");
-                                                                }
-                                                            }
-                                                            let frame_difficulty = exec_frame.header.as_ref()
-                                                                .map(|h| h.difficulty)
-                                                                .unwrap_or(0);
-                                                            pl_for_recv.set_prover_root_verified_frame(exec_num);
-                                                            if !archive_mode_recv {
-                                                                match pl_for_recv.evaluate(
-                                                                    exec_num,
-                                                                    frame_difficulty as u64,
-                                                                    pr_for_recv.as_ref(),
-                                                                    wm_for_recv.as_ref(),
-                                                                ) {
-                                                                    Ok(actions) => {
-                                                                        for action in actions {
-                                                                            info!(frame = exec_num, ?action, "prover lifecycle action");
-                                                                            pp_for_recv.dispatch(action);
-                                                                        }
-                                                                    }
-                                                                    Err(e) => {
-                                                                        debug!(error = %e, "prover lifecycle evaluation skipped");
-                                                                    }
-                                                                }
-                                                            }
-                                                        }
-                                                        Err(e) => {
-                                                            info!(
-                                                                frame = exec_num,
-                                                                total = frames_received,
-                                                                error = %e,
-                                                                "received GlobalFrame (processing failed)"
-                                                            );
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                            Err(e) => {
-                                                warn!(error = %e, "failed to store frame");
-                                            }
-                                        }
-                                    }
-                                    Err(e) => {
-                                        let prefix = if received.data.len() >= 8 {
-                                            hex::encode(&received.data[..8])
-                                        } else {
-                                            hex::encode(&received.data)
-                                        };
-                                        warn!(
-                                            error = %e,
-                                            bytes = received.data.len(),
-                                            prefix = %prefix,
-                                            "GLOBAL_FRAME decode failed"
-                                        );
-                                    }
-                                }
-                            }
-                            GLOBAL_CONSENSUS => {
-                                consensus_msgs_received += 1;
-                                let current_rank = frames_received;
-                                mc_for_recv.add_message(current_rank, received.data.clone());
-                                // Route inbound consensus messages to the event
-                                // loop handle (populated once activation completes).
-                                if let Some(handle) = ch_for_recv.get() {
-                                    if let Some(tp) = quil_engine::consensus_wire::peek_consensus_type(&received.data) {
-                                        match tp {
-                                            quil_engine::consensus_wire::GLOBAL_PROPOSAL_TYPE => {
-                                                match quil_engine::consensus_wire::GlobalProposal::from_canonical_bytes(&received.data) {
-                                                    Ok(wire) => {
-                                                        match quil_engine::consensus_types::wire_proposal_to_signed(wire) {
-                                                            Ok((sp, qc, _tc)) => {
-                                                                handle.submit_quorum_certificate(qc);
-                                                                // Skip pre-submitting the
-                                                                // proposal's
-                                                                // `previous_rank_timeout_certificate`.
-                                                                // Same hazard as the TimeoutState
-                                                                // path below: an unvalidated TC
-                                                                // would land in the pacemaker's
-                                                                // newest-TC tracker and be
-                                                                // embedded into our own next
-                                                                // outgoing timeout. Validation
-                                                                // happens later in
-                                                                // `validate_proposal` →
-                                                                // `validate_timeout_certificate`,
-                                                                // and a real TC will surface via
-                                                                // the local timeout aggregator's
-                                                                // `on_tc_created` callback once
-                                                                // enough peer timeouts arrive.
-                                                                // Feed into the rank's vote collector
-                                                                // so the proposer's self-vote counts
-                                                                // toward quorum and subsequent
-                                                                // standalone votes get verified.
-                                                                if let Some(agg) = va_for_recv.get() {
-                                                                    agg.handle_proposal(&sp);
-                                                                }
-                                                                let h = handle.clone();
-                                                                tokio::spawn(async move {
-                                                                    h.submit_proposal(sp).await;
-                                                                });
-                                                            }
-                                                            Err(e) => warn!(error = %e, "GlobalProposal bridge failed"),
-                                                        }
-                                                    }
-                                                    Err(e) => warn!(error = %e, "GlobalProposal decode failed"),
-                                                }
-                                            }
-                                            quil_engine::consensus_wire::PROPOSAL_VOTE_TYPE => {
-                                                // Route standalone votes into the per-rank aggregator.
-                                                // On reaching quorum, the aggregator's
-                                                // OnQuorumCertificateCreated callback fires
-                                                // `handle.submit_quorum_certificate`.
-                                                match quil_engine::consensus_wire::ProposalVote::from_canonical_bytes(&received.data) {
-                                                    Ok(wire) => {
-                                                        if let Some(agg) = va_for_recv.get() {
-                                                            let gv = quil_engine::vote_aggregation::wire_vote_to_global_vote(wire);
-                                                            agg.handle_vote(gv);
-                                                        }
-                                                    }
-                                                    Err(e) => warn!(error = %e, "ProposalVote decode failed"),
-                                                }
-                                            }
-                                            quil_engine::consensus_wire::TIMEOUT_STATE_TYPE => {
-                                                match quil_engine::consensus_wire::TimeoutState::from_canonical_bytes(&received.data) {
-                                                    Ok(ts) => {
-                                                        // Fast-forward the newest-QC tracker.
-                                                        // Safe: bad QCs fail later validation.
-                                                        let qc_for_handle = ts.latest_quorum_certificate.clone().into_trait_object();
-                                                        handle.submit_quorum_certificate(qc_for_handle);
-                                                        // DO NOT auto-submit the embedded
-                                                        // `prior_rank_timeout_certificate` —
-                                                        // a malformed TC would land in our
-                                                        // pacemaker's newest-TC tracker and
-                                                        // get embedded into our next timeout,
-                                                        // which peers then reject. Outgoing
-                                                        // TCs source from clock store
-                                                        // (previously validated). Local
-                                                        // aggregation forms a valid TC once
-                                                        // peer timeouts arrive.
-                                                        if let Some(agg) = ta_for_recv.get() {
-                                                            let typed = quil_engine::timeout_aggregation::wire_timeout_to_typed(ts);
-                                                            agg.handle_timeout(typed);
-                                                        }
-                                                    }
-                                                    Err(e) => warn!(error = %e, "TimeoutState decode failed"),
-                                                }
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                }
-                            }
-                            GLOBAL_PROVER => {
-                                prover_msgs_received += 1;
-                                let current_rank = frames_received;
-                                mc_for_recv.add_message(current_rank, received.data.clone());
-                            }
-                            GLOBAL_ALERT => {
-                                // Mirrors Go's alert validation at
-                                // `message_validation.go:641-657` +
-                                // `handleAlertMessage` at
-                                // `message_processors.go:810`. The
-                                // alert's Ed448 signature must verify
-                                // against the configured `alertKey`
-                                // with domain `"GLOBAL_ALERT" || message`.
-                                // Canonical format:
-                                //   [u32 type=0x0911][u32 msg_len][msg][u32 sig_len][sig]
-                                if alert_pubkey_for_recv.is_empty() || alert_pubkey_for_recv.len() != 57 {
-                                    debug!("GLOBAL_ALERT received but no valid alertKey configured — dropping");
-                                } else if received.data.len() >= 12 {
-                                    let d = &received.data;
-                                    let mut c = 4usize; // skip type prefix
-                                    let msg_len = u32::from_be_bytes(d[c..c+4].try_into().unwrap_or([0;4])) as usize;
-                                    c += 4;
-                                    if msg_len <= 1000 && c + msg_len + 4 <= d.len() {
-                                        let msg = &d[c..c+msg_len];
-                                        c += msg_len;
-                                        let sig_len = u32::from_be_bytes(d[c..c+4].try_into().unwrap_or([0;4])) as usize;
-                                        c += 4;
-                                        if sig_len == 114 && c + sig_len <= d.len() {
-                                            let sig = &d[c..c+sig_len];
-                                            let mut signed = Vec::with_capacity(b"GLOBAL_ALERT".len() + msg_len);
-                                            signed.extend_from_slice(b"GLOBAL_ALERT");
-                                            signed.extend_from_slice(msg);
-                                            if quil_crypto::ed448_verify(&alert_pubkey_for_recv, &signed, sig) {
-                                                let msg_str = String::from_utf8_lossy(msg);
-                                                warn!(message = %msg_str, "GLOBAL ALERT (verified)");
-                                                coverage_for_recv.emit_alert(&msg_str);
-                                            } else {
-                                                debug!("GLOBAL ALERT rejected — signature invalid");
-                                            }
-                                        } else {
-                                            debug!(sig_len, "GLOBAL ALERT rejected — bad signature length");
-                                        }
-                                    } else {
-                                        debug!(msg_len, "GLOBAL ALERT rejected — bad message length");
-                                    }
-                                }
-                            }
-                            _ => {
-                                // Per-shard routing: if the bitmask matches one
-                                // of the four bitmasks for an active shard
-                                // engine on this node, forward the bytes to
-                                // the engine via its `AppEngineHandle`. The
-                                // worker thread loops own messages back via
-                                // `app_handle.send(...)`, so we must not also
-                                // route self-published messages here — the
-                                // BlossomSub mesh already drops self-echoes.
-                                let bm = received.bitmask.as_slice();
-                                // Snapshot the active filter set under the read
-                                // lock, then drop it before doing per-handle
-                                // sends (the channel is bounded; sends are
-                                // try_send).
-                                let entries: Vec<(Vec<u8>, quil_engine::app_engine::AppEngineHandle)> = {
-                                    let map = shard_engines_for_recv.read();
-                                    map.iter()
-                                        .map(|(f, h)| (f.clone(), h.clone()))
-                                        .collect()
-                                };
-                                let mut routed = false;
-                                for (filter, handle) in &entries {
-                                    if bm == quil_engine::bitmasks::shard_consensus_bitmask(filter).as_slice() {
-                                        handle.send(quil_engine::app_engine::AppEngineMessage::Consensus(received.data.clone()));
-                                        routed = true;
-                                        break;
-                                    }
-                                    if bm == quil_engine::bitmasks::shard_frame_bitmask(filter).as_slice() {
-                                        handle.send(quil_engine::app_engine::AppEngineMessage::Frame(received.data.clone()));
-                                        routed = true;
-                                        break;
-                                    }
-                                    if bm == quil_engine::bitmasks::shard_prover_bitmask(filter).as_slice() {
-                                        handle.send(quil_engine::app_engine::AppEngineMessage::Prover(received.data.clone()));
-                                        routed = true;
-                                        break;
-                                    }
-                                    if bm == quil_engine::bitmasks::shard_dispatch_bitmask(filter).as_slice() {
-                                        handle.send(quil_engine::app_engine::AppEngineMessage::Dispatch(received.data.clone()));
-                                        routed = true;
-                                        break;
-                                    }
-                                }
-                                if !routed {
-                                    // Non-shard traffic (e.g. mesh relay) — no local handler.
-                                }
-                            }
-                            }
-                        }
-                        None => {
-                            info!("message channel closed");
-                            break;
-                        }
-                    }
-                }
-                _ = recv_token.cancelled() => {
-                    break;
-                }
-            }
-        }
-        info!(
-            frames = frames_received,
-            peer_infos = peer_infos_received,
-            "message receiver stopped"
-        );
+    master_node::message_loop::spawn(master_node::message_loop::MessageLoopArgs {
+        clock_store: clock_store.clone(),
+        exec_manager: exec_manager.clone(),
+        token: token.clone(),
+        msg_rx,
+        consensus_loopback_rx,
+        global_msg_tx: global_msg_tx.clone(),
+        archive_pool: archive_pool.clone(),
+        mtls_seed,
+        hg_store: hg_store.clone(),
+        frame_validator,
+        message_collector: message_collector.clone(),
+        coverage_monitor: coverage_monitor.clone(),
+        worker_allocator: worker_allocator.clone(),
+        prover_pipeline: prover_pipeline.clone(),
+        consensus_handle: consensus_handle.clone(),
+        vote_aggregator: vote_aggregator.clone(),
+        timeout_aggregator: timeout_aggregator.clone(),
+        peer_info_cache: peer_info_cache.clone(),
+        shard_engines: shard_engines.clone(),
+        signer_registry: signer_registry.clone(),
+        current_frame: current_frame.clone(),
+        last_global_head_frame: last_global_head_frame.clone(),
+        genesis_archive_peer_ids: genesis_archive_peer_ids.clone(),
+        genesis_prover_addrs: genesis_prover_addrs.clone(),
+        alert_pubkey: hex::decode(&config.engine.alert_key).unwrap_or_default(),
+        network,
+        archive_mode,
+        prover_lifecycle: prover_lifecycle.clone(),
+        prover_registry: prover_registry.clone(),
+        worker_manager: worker_manager.clone(),
+        prover_address,
+        p2p_handle: p2p_handle.clone(),
     });
 
     // ---------------------------------------------------------------
