+++ crates/quil-node/src/master_node/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod allocator_and_lifecycle;
 pub(crate) mod engines;
 pub(crate) mod frame_pipeline;
 pub(crate) mod keys;
+pub(crate) mod message_loop;
 pub(crate) mod networking;
 pub(crate) mod peer_info_publisher;
 pub(crate) mod runtime_state;
